requirements.txt
fastapi
google-generativeai
openai
psycopg[binary]
python-multipart
python-dotenv
redis
uvicorn[standard]
